Add healthz server
A healthz server is added here to indicate when the controller is ready.
There are a few caveats here to handle multiple routers being started in
a single application.

First, only one healthz server is created per application. The healthz
server can be configured to check the health of multiple routers, but
the first router created will determine to which port the server binds.
The default router will have a port of 8888. However, passing a
non-positive port to the New function will result in the router not
being monitored by the healthz server. This is useful for situations
when many routers are started and stopped in a single application. If
these types of routers exist in an application, it is best to turn off
healthz checks for them.

Signed-off-by: Donnie Adams <donnie@acorn.io>