Fix the unregister matcher
The unregister matcher was too broad; we should only be removing
matchers that match the object exactly, not matchers that match via
listing. This change accomplishes that.

Additionally, since there is only one type implementing the matcher
interface, the interface is removed and the type used explicitly.

Finally, the unregistering the triggers also triggers matchers that
match the deleted object. This means that calling Unregister and Trigger
is no longer necessary: one or the other is called.

Signed-off-by: Donnie Adams <donnie@acorn.io>