Add multi-client/cache support
Previously, if a user wanted to create a router that supported multiple
different API endpoints or have a router watch an object in one API and
create object(s) in another API, then the user would have to create two
routers or pass around a client.

Now, the router configuration can take "SpecialConfigs" that it will use
when making requests for specific groups. So, when making a request for
a special group, the corresponding client or cache will be used. For
groups that don't have such special configuration, the default client is
used. This means that all of baaah's functionality can be used across
API endpoints with a single router.

Signed-off-by: Donnie Adams <donnie@acorn.io>